% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifitter-package.R
\docType{package}
\name{epifitter-package}
\alias{epifitter}
\alias{epifitter-package}
\title{epifitter: Analysis and simulation of plant disease progress curves}
\description{
Internal package documentation used to generate the package namespace.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlvesKS/epifitter}
  \item \url{https://alvesks.github.io/epifitter/}
  \item Report bugs at \url{https://github.com/AlvesKS/epifitter/issues}
}

}
\author{
\strong{Maintainer}: Kaique dos S. Alves \email{kaiquedsalves@gmail.com} (\href{https://orcid.org/0000-0001-9187-0252}{ORCID})

Authors:
\itemize{
  \item Emerson M. Del Ponte \email{delponte@ufv.br} (\href{https://orcid.org/0000-0003-4398-409X}{ORCID})
  \item Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
}

}
\keyword{internal}
