% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expo_fun.R
\name{expo_fun}
\alias{expo_fun}
\title{Exponential model differential equation}
\usage{
expo_fun(t, y, par)
}
\arguments{
\item{t}{Time.}

\item{y}{State variable.}

\item{par}{Model parameters.}
}
\value{
A list containing the rate of change.
}
\description{
Internal helper used by the simulation functions to solve the exponential
epidemic model with `deSolve::ode()`.
}
