% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompi_fun.R
\name{gompi_fun}
\alias{gompi_fun}
\title{Gompertz model differential equation}
\usage{
gompi_fun(t, y, par)
}
\arguments{
\item{t}{Time.}

\item{y}{State variable.}

\item{par}{Model parameters.}
}
\value{
A list containing the rate of change.
}
\description{
Internal helper used by the simulation functions to solve the Gompertz
epidemic model with `deSolve::ode()`.
}
