% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_fun.R
\name{mono_fun}
\alias{mono_fun}
\title{Monomolecular model differential equation}
\usage{
mono_fun(t, y, par)
}
\arguments{
\item{t}{Time.}

\item{y}{State variable.}

\item{par}{Model parameters.}
}
\value{
A list containing the rate of change.
}
\description{
Internal helper used by the simulation functions to solve the monomolecular
epidemic model with `deSolve::ode()`.
}
