% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fit_lin.R
\name{print.fit_lin}
\alias{print.fit_lin}
\title{Print fitted model summaries}
\usage{
\method{print}{fit_lin}(x, ...)
}
\arguments{
\item{x}{An object produced by [fit_lin()] or [fit_nlin()].}

\item{...}{Further arguments passed to [print()].}
}
\description{
Print method for objects returned by [fit_lin()] and compatible fitters.
}
