% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fit_nlin2.R
\name{print.fit_nlin2}
\alias{print.fit_nlin2}
\title{Print fitted model summaries with asymptote estimates}
\usage{
\method{print}{fit_nlin2}(x, ...)
}
\arguments{
\item{x}{An object produced by [fit_nlin2()].}

\item{...}{Further arguments passed to [print()].}
}
\description{
Print method for objects returned by [fit_nlin2()].
}
