\name{expo_fun}
\alias{expo_fun}
\title{Function for Exponential model}
\usage{
expo_fun(t, y, par)
}
\description{
Base function for the Exponential model. This function is used in the Exponential model simulation function \code{sim_exponential()}
}
\arguments{
\item{t}{Vector of time}
\item{y}{Vector of disease intensity}
\item{par}{List of parameters}
}

