\name{gompi_fun}
\alias{gompi_fun}
\title{Function for Gompertz model}
\usage{
gompi_fun(t, y, par)
}
\description{
Base function for the Gompertz model. This function is used in the Gompertz model simulation function \code{sim_gompertz()}
}
\arguments{
\item{t}{Vector of time}
\item{y}{Vector of disease intensity}
\item{par}{List of parameters}
}
