\name{logi_fun}
\alias{logi_fun}
\title{Function for logistic model}
\usage{
logi_fun(t, y, par)
}
\description{
Base function for the Logistic model. This function is used in the Logistic model simulation function \code{sim_logistic()}
}
\arguments{
\item{t}{Vector of time}
\item{y}{Vector of disease intensity}
\item{par}{List of parameters}
}

