\name{mono_fun}
\alias{mono_fun}
\title{Function for Monomolecular model}
\usage{
mono_fun(t, y, par)
}
\description{
Base function for the Monomolecular model. This function is used in the Monomolecular model simulation function \code{sim_monomolecular()}
}
\arguments{
\item{t}{Vector of time}
\item{y}{Vector of disease intensity}
\item{par}{List of parameters}
}
